+++ src/tool/index.ts
@@ -1,7 +1,7 @@
 export { default as toolConfig } from './tool.config';
 export { templateBaseVersion } from './tool.config';
 export { templateMetadata, getPublicSiteUrl } from './template-metadata';
-export { textCaseTool } from './tool-definition';
+export { textCaseTool, ALL_VALID_MODES } from './tool-definition';
 export { ToolCanvas } from './components/tool-canvas';
 export { ToolToolbar } from './components/tool-toolbar';
 export { ToolSidebar } from './components/tool-sidebar';